Moved patternverifier-related things into their own folder. Renamed IVerifier to IPatternVerifier
