+++ src/main/java/PatternVerifiers/SingletonVerifier.java
@@ -1,10 +1,10 @@
-package base;
+package PatternVerifiers;
 
 import org.gradle.internal.impldep.org.apache.commons.lang.NotImplementedException;
 
 import com.github.javaparser.ast.CompilationUnit;
 
-public class SingletonVerifier implements IVerifier {
+public class SingletonVerifier implements IPatternVerifier {
 
     @Override
     public boolean verify(CompilationUnit cu) {
